Converted project into modules, created session object file
